pyproject.toml
[tool.poetry]
name = "landbot-test-bot"
version = "0.1.0"
description = "Landbot Test Bot"
authors = ["Pablo Cabezas <pabcabsal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13"
uvicorn = "^0.11"
jinja2 = "^2.11"
pyyaml = "^5.3"
ujson = "^3.0"
itsdangerous = "^1.1"
python-dotenv = "^0.14"
tortoise-orm = {extras = ["accel"], version = "^0.16.13"}

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-asyncio = "^0.12.0"
async-asgi-testclient = "^1.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


